move Result impl into root
